--- lib/angular-rails-templates.rb
@@ -1,7 +1,0 @@
-require 'angular-rails-templates/engine'
-
-module AngularRailsTemplates
-  autoload :HamlProcessor, 'angular-rails-templates/haml_processor'
-  autoload :Processor, 'angular-rails-templates/processor'
-  autoload :VERSION,  'angular-rails-templates/version'
-end
