+++ lib/angular-rails4-templates.rb
@@ -1,0 +1,7 @@
+require 'angular-rails4-templates/engine'
+
+module AngularRails4Templates
+  autoload :HamlProcessor, 'angular-rails4-templates/haml_processor'
+  autoload :Processor, 'angular-rails4-templates/processor'
+  autoload :VERSION,  'angular-rails4-templates/version'
+end
