+++ lib/angular-rails4-templates/compact_javascript_escape.rb
