+++ lib/angular-rails4-templates/engine.rb
