+++ lib/angular-rails4-templates/haml_processor.rb
