+++ lib/angular-rails4-templates/javascript_template.js.erb
@@ -4,3 +4,4 @@
 angular.module("<%= angular_module %>").run(["$templateCache", function($templateCache) {
   $templateCache.put("<%= angular_template_name %>", <%= html %>)
 }]);
+
