+++ lib/angular-rails4-templates/processor.rb
