+++ lib/angular-rails4-templates/template.rb
@@ -51,3 +51,4 @@ module AngularRails4Templates
     end
   end
 end
+
