+++ lib/angular-rails4-templates/version.rb
