+++ test/dummy/app/assets/javascripts/application.js
@@ -1,3 +1,3 @@
-//= require angular-rails-templates
+//= require angular-rails4-templates
 //= require_tree .
 //= require_tree ../templates
